face_verification/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
deepface==0.0.95
numpy==2.2.6
opencv-python==4.12.0.88
Pillow==11.3.0
